requirements.txt
torch
torchvision
pillow
opencv-python
numpy
pathlib
tqdm
segmentation-models-pytorch
wandb
scikit-learn
